Forward query allocation and aggregate the results
This will forward the query allocation when received
between the connected listeners and aggregates the
results of each listeners. This helps to indicate
the upstream elements which metas are supported.
For example it avoids v4l2src buffer copy when the
downstream elements doesn't report that video meta
is supported and the pitch is different than width.