Don't consider old filters on new caps queries
While this is not typically a problem, there are some cases where the
filter included in the caps query may vary drastically in an attempt
to find the most appropriate set of caps. Previously, the last set of
negtiated caps, including the filter, were being stored to be used in
the next query. However, when the filter changed drastically, the
intersection was empty and caps were never negotiated.